+++ src/main/java/Dodo.java
@@ -1,0 +1,27 @@
+import java.util.Scanner;
+public class Dodo {
+    public static void main(String[] args) {
+        Scanner scanner = new Scanner(System.in);
+        String divider = "________________________";
+        String logo =
+                  " ____        _        \n"
+                + "|  _ \\ _   _| | _____ \n"
+                + "| | | | | | | |/ / _ \\\n"
+                + "| |_| | |_| |   <  __/\n"
+                + "|____/ \\__,_|_|\\_\\___|\n";
+        System.out.println("Hello from\n" + logo);
+        System.out.println("What can I do for you?");
+        System.out.println(divider);
+        while (scanner.hasNext()) {
+            String input = scanner.next();
+            if (input.equals("bye")) {
+                System.out.println("Bye. Hope to see you again soon!");
+                System.out.println(divider);
+                break;
+            }
+            System.out.println(input);
+            System.out.println(divider);
+        }
+    }
+
+}
